--- docs/source/api/constants.rst
@@ -1,12 +1,0 @@
-Constants and Configuration
-===========================
-
-This module contains constants and default values used throughout HBAT.
-
-Constants Module
-----------------
-
-.. automodule:: hbat.constants
-   :members:
-   :undoc-members:
-   :show-inheritance:
