+++ docs/source/api/constants/atomic_data.rst
@@ -1,0 +1,23 @@
+Atomic Data
+===========
+
+Contains atomic properties and constants for elements commonly found in PDB structures.
+
+Module Overview
+---------------
+
+.. automodule:: hbat.constants.atomic_data
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Classes
+-------
+
+Atomic Data
+~~~~~~~~~~~
+
+.. autoclass:: hbat.constants.atomic_data.AtomicData
+   :members:
+   :undoc-members:
+   :show-inheritance:
