+++ docs/source/api/constants/index.rst
@@ -1,0 +1,16 @@
+Constants
+---------
+
+The constants package contains all constants, default values, and configuration parameters used throughout HBAT. The package is organized into specialized modules for better maintainability and logical grouping.
+
+Modules
+-------
+
+.. toctree::
+   :maxdepth: 2
+
+   parameters
+   atomic_data
+   pdb_constants
+   misc
+   app
