+++ docs/source/api/constants/misc.rst
@@ -1,0 +1,39 @@
+Miscellaneous Constants
+=======================
+
+Contains GUI defaults, file formats, and other miscellaneous constants.
+
+Module Overview
+---------------
+
+.. automodule:: hbat.constants.misc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Classes
+-------
+
+GUI Defaults
+~~~~~~~~~~~~
+
+.. autoclass:: hbat.constants.misc.GUIDefaults
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Vector Defaults
+~~~~~~~~~~~~~~~
+
+.. autoclass:: hbat.constants.misc.VectorDefaults
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+File Formats
+~~~~~~~~~~~~
+
+.. autoclass:: hbat.constants.misc.FileFormats
+   :members:
+   :undoc-members:
+   :show-inheritance:
