+++ hbat/utilities/__init__.py
@@ -1,0 +1,9 @@
+"""
+Utilities Module
+
+This module contains utility functions that are used across different parts of HBAT.
+"""
+
+from .atom_utils import get_element_from_pdb_atom, pdb_atom_to_element
+
+__all__ = ["get_element_from_pdb_atom", "pdb_atom_to_element"]
